// src/app/core/models/participants.interface.ts
export interface IParticipantData {
  username?: string | null;
  password?: string | null;
  fullName?: string | null;
  instituteName?: string | null;
  country?: string | null;
  city?: string | null;
  continent?: string | null;
  phoneNumber?: string | null;
  faxNumber?: string | null;
  poBox?: string | null;
  registrationNumber?: string | null;
  certificateNumber?: string | null;
  academicNumber?: string | null;
  sector?: string | null;
  websiteUrl?: string | null;
  addressLine?: string | null;
  zipCode?: string | null;
  bio?: string | null;
  participantType?: string | null;
  domains?: any[];
  profilePicBase64?: string | null;
}